chore: move file
使用 views 替换 pages
