+++ src/router/index.js
@@ -1,6 +1,6 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
-import index from '@/pages/index/index'
+import index from '@/views'
 
 Vue.use(VueRouter)
