+++ src/views/index.vue
@@ -6,8 +6,9 @@
 </template>
 
 <script>
-import Loading from '../../components/Loading'
-import CodemirrorEditor from './view/CodemirrorEditor'
+import Loading from '../components/Loading'
+import CodemirrorEditor from './CodemirrorEditor'
+
 export default {
   name: `App`,
   components: {
@@ -32,10 +33,12 @@ export default {
 .fade-leave-to {
   opacity: 0;
 }
+
 .fade-enter-to,
 .fade-leave {
   opacity: 1;
 }
+
 .fade-enter-active,
 .fade-leave-active {
   transition: all 1s;
